Complex and function-=pointer types now tested to work with var arg functions

+++ pyllars/pyllars/pyllars_utils.hpp
@@ -561,7 +561,7 @@ namespace __pyllars_internal {
     template <typename T1, typename T2>
     inline size_t offset_of(T1 T2::*member) {
         T2 object {};
-        return size_t(&(object.*member)) - size_t(&object);
+        return size_t((long long)(&(object.*member)) - (long long)(&object));
     }
 }
